requirements.txt
attrs==24.3.0
aws-cdk-lib==2.179.0
aws-cdk.asset-awscli-v1==2.2.224
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==39.2.20
cattrs==24.1.2
constructs==10.4.2
importlib_resources==6.5.2
jsii==1.106.0
publication==0.0.3
python-dateutil==2.9.0.post0
six==1.17.0
typeguard==2.13.3
typing_extensions==4.12.2
